Revert changes to BinaryOperation
These are no longer required now that ApplyFloodWorkspace has been moved back to Algorithms

Re #23012
